My code for 1626C in Python
Solved 69 minutes into Educational Codeforces Round 121 (Rated for Div 2) with AC first try. However, the site crashed about 70 minutes into the round and it will most likely be voided. It was a sheer coincidence that the site crashed just as I was trying to load the statement for problem D.